requirements.txt
alembic==0.8.8
click==6.6
Flask==0.11.1
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
psycopg2==2.6.1
python-editor==1.0.1
python-firebase==1.2
requests==2.11.1
SQLAlchemy==1.1.1
Werkzeug==0.11.11
WTForms==2.1
